refactor: move org-scoped pages into their own /app application shell
Per the follow-up review on #691/#702, the org-scoped organizer pages
(dashboard/engagements/members/settings) were still just tab content nested
one level inside the same public-site shell (AppLayout), with the org
switcher mounted in the global header on every route regardless of whether
it was relevant there. This restructures them into a genuinely separate
application context:

- New /app/{org-slug}/{dashboard,engagements,members,settings} routes, each
  a real page (frontend/src/pages/app/*) rendered through a new
  OrgAppLayout - its own header/switcher/tab-nav/footer, not AppLayout.
  frontend/src/pages/OrganizationOverviewPage.tsx (the old tab-switched
  monolith) is deleted; its logic is split across the four new pages.
- OrganizationSwitcher moved out of the global Header (both desktop and
  mobile) into the /app shell only, and was redesigned rather than just
  relocated: each row is now a single click target (the old duplicate
  name-click + dashboard-icon-button pair collapsed into one), switching
  orgs preserves whichever tab you're on instead of always resetting to the
  dashboard, and it now lists only organizations the user actually
  organizes (see backend change below) instead of every org they're a
  member of.
- ProfileOverviewPage gained a "Your organizations" entry point (the only
  way into /app now that the header no longer carries it) and creating an
  org from either the switcher or the profile page navigates straight into
  its new /app dashboard.
- Public /organizations/{id} (org profile) is untouched - it stays on the
  Main Page, unauthenticated, addressed by slug. Editing that same data
  moved into /app/{slug}/settings, consistent with editing being an
  organizer action.
- Old /organizations/{id}/dashboard|settings|engagements URLs redirect into
  the new /app equivalents for pre-restructure bookmarks/links.
- Removed the "active org" cookie entirely - with the switch itself always
  explicit now (a full list on the profile page, or a real navigation on
  switch), there's no separate persisted "default org" concept left to get
  out of sync.

Backend: GetOrganizations (feeding the switcher and the profile-page entry
point) used to return every Keycloak org the user was a member of,
regardless of role - a plain member would see an org they can't actually
open (GetOrganizationDetails requires Organizer). Added
IApplicationDbContext.GetOrganizerOrganizationsAsync, sourced from the
organization_membership table added in bea80bc/4caa139, and pointed the
handler at it instead - which also let the now-unused
IKeycloakOrganizationService.GetUserOrganizationsAsync and its backing
Keycloak call be removed.

Test plan:
- dotnet build - passes, 0 warnings
- dotnet test tests/Application.UnitTests - 242/242 passed (added
  GetOrganizationsQueryHandlerTests covering the organizer-only filter)
- dotnet test tests/ArchitectureTests - 247/247 passed
- pnpm check (tsc), pnpm lint, pnpm format:write - all clean
- Updated backend/tests/VisualTests/{NavigationTests,OrganizationTests,
  OrganizationEngagementsTabTests,AccessibilityTests,
  SharedFormClassesTests,VolunteerOpportunityTests,
  CheckInPinOrganizerSetTests}.cs for the new /app routes and the
  switcher's new location; added OrgAppRestructureTests.cs (header switcher
  absence outside /app, profile entry point, legacy URL redirect) and three
  new AccessibilityTests.cs scans (Dashboard/Engagements/Members pages had
  no axe coverage under the old tab-based single page either)
- /self-review fan-out: architecture-check (clean - new
  GetOrganizerOrganizationsAsync fits the existing IApplicationDbContext
  cross-aggregate query pattern, no layer violations; flagged a stale
  backend/CLAUDE.md doc reference to the removed Keycloak method, fixed),
  ef-migration-check (clean - pure query addition against
  already-migrated tables, no schema change), i18n-check (clean - en/de
  key parity confirmed, all new t() calls in the ported pages resolve),
  a11y-check (flagged Dashboard/Members/Engagements pages had zero axe
  coverage and aria-current inconsistency vs the tab nav - both fixed;
  flagged the calendar event color-picker dialog and SlugGenerator
  reserved-word collisions with the new /app path segments as lower-value
  follow-ups, spawned as separate tasks rather than folded in here)
- IntegrationTests and VisualTests need Testcontainers/Aspire, not reliably
  available in this sandbox - rely on CI's dotnet.yml for those, plus the
  live-verification pass to follow

Addresses the restructuring requested in the #702 review, on top of the
already-merged-to-this-branch organizer-lockout backfill fix (0fb737d).
