+++ backend/src/Application/Organizations/GetOrganizations/v1/GetOrganizationsQueryHandler.cs
@@ -1,11 +1,10 @@
-using Application.Common.Keycloak;
 using Application.Common.Messaging;
 using Application.Common.Persistence;
+using Domain.Users;
 
 namespace Application.Organizations.GetOrganizations.v1;
 
 internal sealed class GetOrganizationsQueryHandler(
-	IKeycloakOrganizationService keycloakOrganizationService,
 	IApplicationDbContext dbContext)
 	: IQueryHandler<GetOrganizationsQuery, IReadOnlyList<OrganizationSummaryDto>>
 {
@@ -13,14 +12,11 @@ internal sealed class GetOrganizationsQueryHandler(
 		GetOrganizationsQuery request,
 		CancellationToken cancellationToken = default)
 	{
-		var organizations = await keycloakOrganizationService.GetUserOrganizationsAsync(
-			request.UserId, cancellationToken);
-
-		var slugs = await dbContext.GetOrganizationSlugsAsync(
-			organizations.Select(o => o.Id).ToList(), cancellationToken);
+		var organizations = await dbContext.GetOrganizerOrganizationsAsync(
+			new UserId(request.UserId), cancellationToken);
 
 		return organizations
-			.Select(o => new OrganizationSummaryDto(o.Id, o.Name, slugs.GetValueOrDefault(o.Id)))
+			.Select(o => new OrganizationSummaryDto(o.Id.Value, o.Name, o.Slug))
 			.ToList();
 	}
 }
